firmware_UpdateKernelSubkeyVersion: Update for vboot2
This patch updates the firmware_UpdateKernelSubkeyVersion test to work
on vboot2 devices (which has somehow been missed until now since that
test isn't actually part of any commonly used FAFT suite for some
reason).

Since vboot2 doesn't copy firmware B back to A on bootok after a
successful update, we should read the updated kernel subkey version from
slot B (which would just be the same on vboot1). We also need to account
for not switching back to slot A automatically after the next reboot.

Also fix an issue introduced by CL:238071, which unintentionally added a
check for fw_try_count=0. This is not true right after an update, but it
becomes true once we run bootok so let's just move the check after that.
(Inline the run_bootok_and_recovery() function to make the whole thing
more readable since it was just a remnant of the old-style FAFT "step"
mechanism and doesn't really make sense as a separate function anymore.)

BUG=chrome-os-partner:42221
TEST=Ran on Jerry.

Change-Id: I7d777ce5e9875ff0ba9abaa926b24e8298483491
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/283517
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>