[autotest] Change DUT status reporting to detect unused hosts.
This adds new handling to status_history.HostJobHistory objects to
return a status of UNUSED when a host is found to be working but
unused in the time range of interest.  This will cause DUTs that
remain locked for an extended period of time to be treated as "not
working" for most purposes.

BUG=chromium:487428
TEST=run `dut-status` against a variety of DUTs.

Change-Id: Ibda0c0ec335750648bffb8a9fe28634cfb1cc8a0
Reviewed-on: https://chromium-review.googlesource.com/284861
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>