Check that video is playing before beginning to capture frames.
I am seeing failures in image comparison tests where we are getting
a high number of duplicate frames. This can be caused if the video
isn't playing, and so chameleon ends up capturing the same frame
over and over again.

Write this check so we can see if video not playing is the cause.

BUG=None.
TEST=Lab will run tests for this.

Change-Id: I06336922a0335c280884809ff512cafd1c64b8eb
Reviewed-on: https://chromium-review.googlesource.com/285290
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>