autotest: Do not fail trying to find the module of wlan0
If the Linux kernel is non-modular and the driver is built into the
kernel, the "driver/module" symlink for that device in sysfs will not
exist.

In that case, fail gracefully by setting the kernel_module field (which
is only used for debugging purposes) to a sensible textual description
indicating that the name of the module is unknown.

BUG=chrome-os-partner:42310
TEST=Ran network_WiFi_SimpleConnect against a machine with a kernel with
  drivers linked into the kernel and modules disabled.
  Ran suite:wifi_matfunc against a Samus in a wifi cage.

Change-Id: If3b822eef5b62238e4094d0f03b6978354a4a1ba
Reviewed-on: https://chromium-review.googlesource.com/284007
Trybot-Ready: Filipe Brandenburger <filbranden@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Filipe Brandenburger <filbranden@chromium.org>
Tested-by: Filipe Brandenburger <filbranden@chromium.org>