autotest: Escape commands run under `adb shell`
This fixes problems such as checking the pid of commands put in
background by echoing the value of "$!". Without the proper escaping,
the value of "$!" would be expanded on the adb_host instead.

Correct an instance in WpaCliProxy running the set_network which tries
to insert literal double quotes and goes to some lengths trying to
"work around" the poor escaping.

The correct way of doing shell escaping would be to use single quotes
(which bypass all expansions and is safe for characters such as "$",
"`", backslash and double quotes) and replace existing instances of "'"
with the expression "'\''", which terminates the current string, inserts
a literal single quote (using backslash for escaping) and then starts a
new string.

However, such a change is too complex for this scope, so I will try to
propose that in a follow up instead.

BUG=chrome-os-partner:42310
TEST=Ran network_WiFi_SimpleConnect through adb successfully.
  Ran suite:wifi_matfunc against a Samus in a wifi cage.

Change-Id: I260a20f2a344754790d509ea801d868965f2fe05
Reviewed-on: https://chromium-review.googlesource.com/285641
Reviewed-by: Filipe Brandenburger <filbranden@chromium.org>
Commit-Queue: Filipe Brandenburger <filbranden@chromium.org>
Tested-by: Filipe Brandenburger <filbranden@chromium.org>