autotest: Refactor code parsing `iw dev` output
The refactor makes it cope with this output (from an Android device):

  $ adb shell iw dev
  phy#0
    Unnamed/non-netdev interface
      wdev 0x2
      addr aa:bb:cc:dd:ee:ff
      type P2P-device
    Interface wlan0
      ifindex 4
      wdev 0x1
      addr aa:bb:cc:dd:ee:ff
      ssid Whatever
      type managed

Before this refactor, interface "wlan0" would be incorrectly associated
with type "P2P-device", since at that point one type and one interface
name were seen.

The correct association for "wlan0" is with type "managed", which
appears under the interface. Types are always expected to be after the
interface, so we should only start looking for a type after seeing an
interface name.

BUG=chrome-os-partner:42310
TEST=Ran network_WiFi_SimpleConnect through adb successfully.
  Ran suite:wifi_matfunc against a Samus in a wifi cage.

Change-Id: Ibdd1199ba27715f5774537e494f69ec6499dbcf9
Reviewed-on: https://chromium-review.googlesource.com/285642
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Filipe Brandenburger <filbranden@chromium.org>
Tested-by: Filipe Brandenburger <filbranden@chromium.org>