autotest: Use `nohup` to put the iw log tail command in background
Otherwise running it through `adb shell` ends up killing it with a
SIGHUP as soon as the shell is terminated.

Also redirect its stdin (to /dev/null) and stderr (dup of stdout) to
prevent the channel from staying open, which would block the script.

BUG=chrome-os-partner:42310
TEST=Ran network_WiFi_SimpleConnect through adb successfully.
  Ran suite:wifi_matfunc against a Samus in a wifi cage.

Change-Id: Ia74f923b0487d6cae99d921876ebdf3d0446120c
Reviewed-on: https://chromium-review.googlesource.com/285643
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Filipe Brandenburger <filbranden@chromium.org>
Tested-by: Filipe Brandenburger <filbranden@chromium.org>