[autotest] Disable provision_AutoUpdate.double for test_that
provision_AutoUpdate.double reads current cros-label from autotest afe
and try to provision the dut to the same build (which ensures that we can
provision out of a build)

This test will fail when running with test_that as it tries to access
cautotest, the default afe server. The test should just be disabled when
running with test_that as there is not afe server in the test_that use
case.

TEST=run provision_AutoUpdate with test_that, ensure it raises TestNA
error.
BUG=chromium:510919

Change-Id: I13b4c026b96097420624a7a03b7eba138f160530
Reviewed-on: https://chromium-review.googlesource.com/286233
Reviewed-by: Mungyung Ryu <mkryu@google.com>
Commit-Queue: Fang Deng <fdeng@chromium.org>
Tested-by: Fang Deng <fdeng@chromium.org>