Freonize ui_AppLauncher and get it running on all devices
ui_AppLauncher checks for unexpected changes in the UI. The code for
keyboard simulation in X is replaced here with keyboard emulation. The
control file has been updated to run on all boards. The logic for
computing the screenshot area now works out the dimensions correctly
for all boards instead of hardcoding it for one board.

BUG=None
TEST=None

Change-Id: I061b54db3ceddbd59fbc07d19810ba2a5115b862
Reviewed-on: https://chromium-review.googlesource.com/286843
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Mussa Kiroga <mussa@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>