[autotest] Sync host's status
For hosts in shards, their status is never synced with master.
Users should go to shard AFE to see the host's status.
This CL syncs the host status via shard heartbeat.

BUG=chromium:510996
TEST=puppylab. Change a host status in stumpy shard manually,
and check if it is synced to the master.
DEPLOY=apache,shard_client

Change-Id: Ic2bdce6aeeea868cbb000a79b6301e4848514be7
Reviewed-on: https://chromium-review.googlesource.com/287615
Reviewed-by: Mungyung Ryu <mkryu@google.com>
Commit-Queue: Mungyung Ryu <mkryu@google.com>
Tested-by: Mungyung Ryu <mkryu@google.com>