Use a Longest Common Subsequence(LCS) to compare chameleon frames.
In our video frame comparison tests we have a list of golden checksums
and a list of test checksums.

Each checksum is computed from each frame. Due to the nature of capturing
frames sometimes we miss a frame. Therefore, either of the two lists could
end up with a missing frame, or what would appear to be an added frame in
the other list.

Currently we are walking the golden checksum list insisting that we find
each checksum in the test checksum list. The test fails when we see a mismatch.

But this mismatch could be because there is one extra or one less checksum, and
the rest of the checksums could continue to be matched.

SO, use LCS to measure how similar the two lists are. If we have less than 10
golden frames that coudln't be found in the test, we will say the test has passed.

In future CLs, a check to make sure the mismatched frames are not consecutive will
be added.

BUG=chromium:484057
TEST=Ran a test 10 times on code without CL. Got a few (false) failures.
     Ran same test with this CL and got no failures.

Change-Id: I5aea72e4cb9ada296e7df365b7108cf969e9c496
Reviewed-on: https://chromium-review.googlesource.com/285854
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>