autotest (wifi): improve packet capture reliability
We've observed that, for very short exchanges, we often miss some
of the transmitted data. In the cases, the output of tcpdump
indicates that frames matched the filter, but were not captured.
E.g.:

  tcpdump: WARNING: monitor0: no IPv4 address assigned
  tcpdump: listening on monitor0, link-type IEEE802_11_RADIO [...]
  0 packets captured
  40 packets received by filter
  0 packets dropped by kernel

Running strace on tcpdump shows that it calls poll(), with a 1 sec
timeout. Some experimentation indicates that sleeping for 1 second,
before terminating tcpdump, ensures that we capture the data we
want. E.g.

  tcpdump -n -i Pmon -s0 -w /tmp/pmon.pcap & sleep 1; \
  send_management_frame -i Qmon -t beacon -n 1; sleep 1; kill $!

So, based on the above, add a short sleep to stop_capture().

Note that I'm not entirely sure why the sleep helps. poll() should
return as soon some FDs are ready. So sleeping for the poll timeout
should just cause us to see poll() report that there are no FDs
are ready.

The answer might be simple. E.g., perhaps the packet filtering
code in the kernel also uses a 1 second timeout, if the capture
buffer hasn't filled.

Alternatively, the answer might be complex. Specifically: the
pcap source (specifically: pcap-linux.c) has a comment about a
bug in TPACKET_V3. The comment and a linked bug suggest a few
more things to try, if this patch doesn't fully resolve the
problem:

	  i) disable TPACKET_V3
	 ii) update to kernel 3.19 or later
	iii) have tcpdump (or our own capture tool) specify a
	     short, but non-zero, timeout
	 iv) use "immediate mode" with libpcap 1.5 or later

Additional references:
[1] https://github.com/the-tcpdump-group/libpcap/issues/335
[2] http://stackoverflow.com/questions/24988182

BUG=chromium:514384
TEST=manual (see above)

Change-Id: I8c4247f4391d7200f25e4bb297bea44ad16b98da
Reviewed-on: https://chromium-review.googlesource.com/288814
Commit-Ready: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>