FAFT: Remove the test updating EC binary inside AP firmware
This test tries to update the EC binary inside the AP firmware
and checks if EC software sync works. However, it requires giving
a new EC binary that is not trivial. As a result, it seldom
be ran, not part belonging to regular FAFT suites.

Actually, this test can be covered by:
 - firmware_SoftwareSync: Verify EC RW able to be updated, via AP
                          copying its EC RW binary to EC.
 - firmware_UpdateFirmwareVersion: Verify AP RW able to be updated,
                                   via the firmware shellback.

CQ-DEPEND=CL:290330
BUG=chromium:516405
TEST=Ran a FAFT test to checked this change not affect others.

Change-Id: I32cfc44854b5f35ade7ab8c752f67bf6a381d815
Reviewed-on: https://chromium-review.googlesource.com/290327
Commit-Ready: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>