autotest (wifi): filter bad frames in get_probe_ssids()
We've found the HiddenScan test test to be flaky with some
hardware combinations. It turns out that we sometimes report
probes for SSIDs which are similar, but not identical, to the
SSID that we set up for the test.

These similar-but-not-identical SSIDs have always appeared
in frames which fail the FCS check. Presumably, either the
client, or the AP, is mangling these frames. (Channel noise
shouldn't be an issue, as these tests are run in a shield
box. And the FCS errors are very frequent.)

To remedy this problem, we have get_probe_ssids() filter out
frames whose FCS values don't match the frame data.

While there:
- Add a |bad_fcs| argument to get_frames(). This argument
  specifies the disposition of frames which fail the FCS check.
  While we could have get_frames() always discard such frames,
  some tests probably won't want that. (E.g., a test of the
  rate control algorithm would want to know how failed
  transmissions affect transmit rates.)
- Rename WLAN_PROBE_REQ_FILTER to WLAN_PROBE_REQ_ACCEPTOR.
  This makes the naming less ambiguous. (We now have both
  positive, and negative filters. "Filter" doesn't tell
  us whether a matching frame will be accepted, or rejected.)
- Update existing tests to specify |bad_fcs='include'|. This
  maintains existing behavior for those tests.
- DarkResumeActiveScans: use ACCEPTOR instead of FILTER.
- TDLSPing: wordsmith some error messages
- TDLSPing: address 'logging-not-lazy' warning from presubmit check.

BUG=chromium:504486
TEST=network_WiFi_HiddenScan [1][2]3[]
TEST=network_WiFi_TDLSPing [2]
TEST=network_WiFi_DarkResumeActiveScans [3]
TEST=network_WiFi_RateControl [2][3]
TEST=network_WiFi_LowInitialBitrates [2][3]
TEST=network_WiFi_VisibleScan [2][3]

[1] 10 iterations, on new hardware
[2] on peppy
[3] on samus

Change-Id: I0c2069bce9e25b70c8d4936265c1d548c814cf18
Reviewed-on: https://chromium-review.googlesource.com/290371
Trybot-Ready: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Mukesh Agrawal <quiche@google.com>