[autotest] Increase USB boot timeout to 5 minutes.
During repair with servo, we boot from USB.  That boot was only
given 150s (2m30s) to complete.  veyron_jerry was requiring 170s
(2m50s) in some cases.  Other boards may also be taking as long.

This increases the allowed time to 300s (5 minutes).  That's really
much longer than we should need, but it's the most expedient thing
to do for now.

BUG=None
TEST=None

Change-Id: I303d96cff1c631c2814d4e2e308774308522df57
Reviewed-on: https://chromium-review.googlesource.com/290352
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>