autotest (wifi): improve reliability of start_capture()
When running the VisibleScan test, we've observed that
the packet capture is sometimes empty. This happens even
in cases where we're reasonably confident that the scan was
initiated after start_capture(). [1] This leads us to believe
that, when start_capture() returns, tcpdump hasn't fully
initialized the packet capture.

To resolve this problem, we have start_capture() wait
until tcpdump has logged the "listening on" message.

[1] This confidence comes from examining timestamped iw event
    logs, alongside i) a timestamped DEBUG log showing when
    tcpdump was launched, and ii) a timestamped DEBUG log
    showing when "iw scan" was launched.

BUG=chromium:514869
TEST=network_WiFi_VisibleScan (10 iterations, new hardware)

Test note: Examining the DEBUG log after this change, we found
that every one of the 10 iterations had two calls to grep,
looking for "listening on". This supports the hypothesis that
start_capture() was returning before tcpdump was fully set up.

Change-Id: Ica7e85b7b6dcbb711622c3f570cc1422dd2c3ece
Reviewed-on: https://chromium-review.googlesource.com/290402
Trybot-Ready: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>