dark_resume_xmlrpc_server: move server constants out
In the lab, we are failing all lucid sleep tests because the
import chain on the server side ends up including stuff that
should only be present on the client side. This breaks the
dependency chain and moves all the relevant constants to
common_lib.

BUG=chromium:485834
TEST=run lucid sleep tests

Change-Id: I082b2b0fba4670fbe148fb353659ff6a34843314
Reviewed-on: https://chromium-review.googlesource.com/291303
Reviewed-by: Samuel Tan <samueltan@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>