[Chameleon] Allow user to specify max number of frames as an arg.
We are capturing frames with chameleon asynchronously. We stop when
we have got over the specified number of frames.

This number has been specified as a configuration and it used during
a test run. However, when we are collecting golden images we might
need to choose a different number without having to change the conf.
file.

BUG=None.
TEST=Ran test locally. Ensured nothing broke.

Change-Id: I1bde45ccad9f3b9f27186c755929a687fd21a159
Reviewed-on: https://chromium-review.googlesource.com/291831
Trybot-Ready: Mussa Kiroga <mussa@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>