faft: Fix lsb-release check for test image
We've been checking for test images in the CHROMEOS_RELEASE_DESCRIPTION
field, which is not actually the correct way to do it (official images
just always contain the word "test" there for no apparent reason, even
if they're actually dev or base images).

This patch makes it check CHROMEOS_RELEASE_TRACK instead which seems to
be consistent on both official and developer-built images for this.

BUG=chromium:519889
TEST=None

Change-Id: I9a83c796e3394bc6ec2a008776eb8bb8ef296739
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/293337
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>