Poll faster for correct number of frames.
The list of test checksums could end up being too long.
This fails the test because those extra checksums are not
in the golden list, hence, the mismatch count is higher.

I believe the problem is that during the lab runs we are
capturing too many frames, perhaps the polling sleeps for too long.

As an experiment, I am reducing the sleep time between polls, making
the loop tighter and hence more likely to perform the check for count
of frames sooner.

BUG=None.
TEST=Seems to repro on lab runs only, so ran a sanity test on falco.

Change-Id: I38175aae941e57d9213dd40f29af2606855ef6d7
Reviewed-on: https://chromium-review.googlesource.com/294154
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>