[autotest] Add a new script that deletes unused labels.
Labels such as 'cros-version', 'fw-version', 'fwrw-version',
and 'fwro-version' are stacking up in the DB, and there are
many such labels that are no longer in use.
This script can be used to delete unused labels.
We will run this script as a cron job in the machines that
run MySQL. Such machines are shards and master DB.

BUG=chromium:380365
TEST=Test with local database.

Change-Id: I0f341c3e5a5977c53b0764efd0442684b525f3fd
Reviewed-on: https://chromium-review.googlesource.com/294164
Trybot-Ready: Mungyung Ryu <mkryu@google.com>
Tested-by: Mungyung Ryu <mkryu@google.com>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Mungyung Ryu <mkryu@google.com>