[autotest] Fix logic of host name for servo_host and chameleon_host
When an IP address is passed in as CrosHost host name, it can not be used
to make Chameleon or Servo host name.  E.g, 192-chameleon.168.1.2 makes
no sense.
Also, in this case, we can not check if it is in lab becase we use
<first part of hostname>.cros.corp.google.com to check if it is in lab.
So we just assume it is not in lab.

BUG=chromium:521443
TEST=run audio_AudioBasicHeadphone test using IP address as Cros host
name, and see the test will not stuck at checking if servo host and
chameleon host are in lab.

Change-Id: I3956eec2631f32f017aa31f23f598c64abf3f5ba
Reviewed-on: https://chromium-review.googlesource.com/295020
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>