FAFT: Use power_key() where appropriate to power on DUT
Some platforms need a longer power button press than others when
turning on. This is particularly true for platforms that do not
have an EC since the EC will hold the power on signal for the host/AP
for as long as needed (which may exceed the SHORT_DELAY used by
power_short_press()). Platforms without an EC need to be able to
define their own power on hold time.

This patch does a few small things to address that:
- Add hold_pwr_button_poweron in FAFT config to be used as the power
  button hold time.
- Rename hold_pwr_button to hold_pwr_button_poweroff to be more
  explicit about its usage.
- In cases where an EC is not used, replace power_short_press() with
  power_key() and pass hold_pwr_button_poweron as its argument.

BUG=chrome-os-partner:44275
BRANCH=firmware-veyron
TEST=firmware_ConsecutiveBoot now passes on mickey (with follow-up
patch to define hold_pwr_button_poweron)

Change-Id: Ie26f254f0cde0ef083f5db3caf51b2c85b622a77
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295186