usb_facade_native: Add USBDeviceDriversManager class and its error class
A USBDeviceDriverManager object manages the drivers of a USB device. It
will provide basic functions to bind and unbind the drivers associated
with the said USB device (to be implemented). It will raise
USBDeviceDriversErrors when binding and unbinding cannot be successfully
performed in scenarios such as when the USB device has not be detected
by the host.

BUG=chromium:514541
TEST=not used yet.

Change-Id: Ib0f9159e0b4a812aad6a14483a99ddee014aedc7
Reviewed-on: https://chromium-review.googlesource.com/294896
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>