autotest: Introduce new sh_quote_word() API for shell commands.
This new API is meant to eventually replace sh_escape().

While the old sh_escape() API works as expected and seems to cover all
the corner cases, I would argue that the new API is:

1) More complete: Because it returns a term that can be directly used as
a shell word (while sh_escape relies on the caller surrounding it with
double quotes.)

2) Cleaner: Most characters are unaffected, including the dollar sign
and backslash and even the double quote.

The only one that gets affected by quoting is the single quote (granted
though, for the single quote it gets pretty ugly, especially if there is
nesting.)

It is also cleaner in that words that require no quoting will be
returned unquoting, with an optional whitelist to control which
characters are totally safe.

Using single quotes for the automated quoting (with sh_quote_word()) is
also nicer as for literal shell commands, the double quote is more
useful, being used for shell code that uses variables, etc. so it is
easier to reason about having all that code use mostly double quotes and
then be fairly recognizable after going through sh_quote_word().

3) More efficient: It only needs a single scan of the original text,
replcaing a single character. (The sh_escape() approach could be made to
do a single scan using regular expressions, but see points #1 and #2
above.)

Add new unit tests to cover the functionality of sh_quote_word. Inherit
from test_sh_escape to ensure the same tests are run on both. Introduce
a new test class to test nested quoting (quote a command, pass it
through "echo" and run the output through another shell.)

BUG=None.
TEST=Ran the base_utils_unittest.py unit tests with no failures.

Change-Id: I8b7fbce0aaf6a7eaa36ee6bc8bb939b915d903fb
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295440
Reviewed-by: Fang Deng <fdeng@chromium.org>