autotest (WiFi): add run-time check for wake-on-WiFi support
Check that the DUT supports wake-on-WiFi in the test body of
wake-on-WiFi tests and throw a TEST_NA failure and end the
test early if the DUT does not support this feature.

This will prevent a confusing test failure from showing up when
test suites including wake-on-WiFi autotests are run directly
on a DUT that does not support wake-on-WiFi using test_that
(which ignores control file depdency labels).

BUG=chrome-os-partner:43973
TEST=Run wifi_lucidsleep test suite on a wificell DUT that does not support
     wake-on-WiFi and verify that all tests pass with a TEST_NA message stating
     that wake-on-WiFi is not supported.

Change-Id: I6b8e1337a4d15823909ca2ef4d9d7ae17646ca8f
Reviewed-on: https://chromium-review.googlesource.com/295427
Tested-by: Samuel Tan <samueltan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Samuel Tan <samueltan@chromium.org>