autotest: Use sh_quote_word() in base_utils.run()
Use the new shell quoting API in the central run() function.

Update test case that depended on the internal quoting format, so that
it now reflects the sh_quote_word() output. Add a comment to note that
specific test depends on the internal quoting function.

It turns out this broke some cases that were passing a list in the
command argument. The old code was working since it used += which is
accepted by a list, but causes problems if args are passed because they
will be in escaped shell syntax. Fix that by making them uniform, always
converting "command" into a string if necessary. (In hindsight, we
should be doing the opposite, but we can try to address that later.)

BUG=None.
TEST=Ran the base_utils_unittest.py unit tests with no failures.

Change-Id: I5ff8f92671583a236ee38d0da90c890bbb096cc8
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295441
Reviewed-by: Christopher Wiley <wiley@chromium.org>