autotest: Cleanup after logging.warn -> logging.warning
There were some small issues which were probably introduced after
replacing logging.warn() calls with logging.warning().

In particular, line wrapping was incorrect in a few places, since line
continuations needed to be indented three extra spaces. In one place,
this bumped the line over 80 columns, so the string had to be rewrapped
as well.

Test cases were still stubbing logging.warn (so logging.warning was not
mocked, producing some noise in the output of test cases), after
correcting it to stub logging.warning, it was clear that some spots were
not marked as expecting a call, so I added those as well.

BUG=None.
TEST=Ran the base_utils_unittest.py unit tests with no failures.
  Also confirmed that the test run did not produce any warning output,
  only the "dots" indicating progress with a final report summary
  ("Ran 91 tests in 2.151s" and the "OK" line.)

Change-Id: I632a59b43d885c7aad8ec2d2083abba08e232279
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295399
Reviewed-by: Christopher Wiley <wiley@chromium.org>