FAFT: Use firmware_screen time to when pressing ctrl-D repeatedly
This patch replaces the time allowed for firmware_DevScreenTimeout
to press ctrl-D from a hard-coded 20sec (increased in CL:240105) to
firmware_screen * 2. Since firmware_screen is currently 10sec by
default, this will not have an impact on most platforms.

The idea for firmware_DevScreenTimeout is to obtain an aggessively
low dev mode boot time value by spamming the AP with ctrl-D every
half second. So firmware_screen in this case acts as a ceiling and
we expect that the AP will be ready to accept ctrl-D earlier.

BUG=chrome-os-partner:43704
BRANCH=firmware-veyron
TEST=firmware_DevScreenTimeout passes more reliably on veyron_mickey

Change-Id: I74fe1060b00a3c027572fc76c44ec54382c3c50f
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295294
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>