Autotest: Add test network_WiFi_RoamDbus
This test configure two APs with the same SSID. The DUT first
connect to AP1. Then we ask shill to send a roam command to
wpa_supplicant through dbus. The roam command is encoded with
the BSSID of AP2. After waiting for a reasonable amount of time,
we verify that DUT is now connected to AP2.

BUG=chromium:523082
TEST=test on a peppy device
CQ-DEPEND=CL:295888,CL:296047,CL:295798

Change-Id: Ib8de6d8341914579a585979edc3372b68288500c
Reviewed-on: https://chromium-review.googlesource.com/296030
Commit-Ready: Ningyuan Wang <nywang@chromium.org>
Tested-by: Ningyuan Wang <nywang@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>