autotest: Fix typo logger -> logging introduced in previous commit
The problem was introduced in I3f06cf3be44430 when we moved the
logging.info calls from the caller to the generic method. My tested
didn't catch it because I was mostly testing the ADBHost case and didn't
really re-run it completely after I moved it out of the "if" case.

BUG=None.
TEST=Re-run it on both ADBHost and not ADBHost cases.

Change-Id: Ic563683e0326ae050014f1d18242e673d3d06ae6
Reviewed-on: https://chromium-review.googlesource.com/295799
Commit-Ready: Filipe Brandenburger <filbranden@chromium.org>
Tested-by: Filipe Brandenburger <filbranden@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>