usb_facade_native: Add _find_usb_device_bus_id to drivers manager class
This is a private helper method that attempts to find, based on the
product name inputted, the bus ID given to the USB device by the host.
This method searches through subdirectories of USB devices in the host
filesystem to find a device with the matching product name, before
extracting its bus ID from the its file path. The bus ID is needed to
bind and unbind drivers associated with the device to control its
availability to the host device.

BUG=chromium:514541
TEST=not used yet.

Change-Id: I0bd6b4beae7df31bdf069a0af678e5fcfdee1209
Reviewed-on: https://chromium-review.googlesource.com/294897
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>