usb_facade_native: Add set_usb_device to drivers manager class
This method allows the user to set the USB device to be managed by the
USBDeviceDriversManager object. In this class, a USB device is
recongized by its product name and its bus id as recognized by the host.
If no USB device is found with the product name given by the user, this
class raises an error.

BUG=chromium:514541
TEST=not used yet.

Change-Id: I1af3aac2b7562bc24216e4e532b8b6eaf38807ef
Reviewed-on: https://chromium-review.googlesource.com/294898
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>