usb_facade_native: Add USBFacadeNative class and its error class
The USBFacadeNative class is implemented to expose USB-specific
functionalities needed on the host side. Right now, the
USBFacadeNative class has a USBDeviceDriversManager object as a
private attribute, which will be used to control the availability/
visibility of the USB device on the host device.

BUG=chromium:514541
TEST=not used yet.

Change-Id: Ie427d5774626efc9fab8485795c1cf68b63741d6
Reviewed-on: https://chromium-review.googlesource.com/296187
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>