[autotest] Collect stats about devserver assignment for provision job.
Collect stats about if a devsever is assigned to a dut in the same subnet or
not, during a provision job.
The stats will be used to evaluate how much cros-lab traffic can be alleviated
if a dut can only use devserver in the same lab.

BUG=chromium:421123
TEST=local

Change-Id: I7ef7e3ea77df99f8d5359b800e4dee4bf7b9eb36
Reviewed-on: https://chromium-review.googlesource.com/296761
Trybot-Ready: Dan Shi <dshi@chromium.org>
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>