[autotest] Convert localhost to host IP if test runs inside container.
FAFT tests in moblab uses servo_host from host attributes. The value could be
set to localhost. When the test runs inside container, autoserv failed to
make servo call as the servo is not localhost for container, rather, it should
be the IP of the moblab. This CL adds such necessary conversion.

BUG=None
TEST=local moblab

Change-Id: I8ea614209a82ae9c8cca1fe927c8c96de4c00879
Reviewed-on: https://chromium-review.googlesource.com/296836
Commit-Ready: Dan Shi <dshi@chromium.org>
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>