FAFT: Non-blocking shell command
Certain shell commands, like "fastboot oem lock",
will not return until detecting an event like the
power button being pushed.  In order for this to be
done in a test, we need to return immediately so that
the servo can emulate the power button push.

BUG=chromium:527484
BRANCH=none
TEST=Tested out with FAFT test to make sure that call
did not wait for return before moving on to next command.

Change-Id: I0c188650fcd4b729e11179563a59d6946f3d70eb
Signed-off-by: Shelley Chen <shchen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/297420
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>